+++ bin/integration-tests/src/main.rs
@@ -206,6 +206,7 @@ impl std::fmt::Debug for TestCase {
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 enum TestCategory {
     Client,
+    ClientService,
     CustomTransaction,
     Fpi,
     NetworkTransaction,
@@ -219,6 +220,7 @@ impl AsRef<str> for TestCategory {
     fn as_ref(&self) -> &str {
         match self {
             TestCategory::Client => "client",
+            TestCategory::ClientService => "client_service",
             TestCategory::CustomTransaction => "custom_transaction",
             TestCategory::Fpi => "fpi",
             TestCategory::NetworkTransaction => "network_transaction",