+++ bin/miden-cli/src/info.rs
@@ -11,7 +11,7 @@ use super::config::CliConfig;
 use crate::commands::account::DEFAULT_ACCOUNT_ID_KEY;
 use crate::errors::CliError;
 
-pub async fn print_client_info<AUTH: TransactionAuthenticator + Sync + 'static>(
+pub async fn print_client_info<AUTH: TransactionAuthenticator + Send + Sync + 'static>(
     client: &Client<AUTH>,
     show_rpc_status: bool,
 ) -> Result<(), CliError> {
@@ -38,7 +38,7 @@ pub async fn print_client_info<AUTH: TransactionAuthenticator + Sync + 'static>(
 
 // HELPERS
 // ================================================================================================
-async fn print_client_stats<AUTH: TransactionAuthenticator + Sync + 'static>(
+async fn print_client_stats<AUTH: TransactionAuthenticator + Send + Sync + 'static>(
     client: &Client<AUTH>,
 ) -> Result<(), CliError> {
     println!("Block number: {}", client.get_sync_height().await?);